:recycle: : moves module feature to its own package
