+++ src/main/java/io/codeka/gaia/controller/IndexController.java
@@ -2,7 +2,7 @@ package io.codeka.gaia.controller;
 
 import io.codeka.gaia.bo.StackState;
 import io.codeka.gaia.repository.StackRepository;
-import io.codeka.gaia.repository.TerraformModuleRepository;
+import io.codeka.gaia.modules.repository.TerraformModuleRepository;
 import io.codeka.gaia.teams.bo.Team;
 import io.codeka.gaia.teams.bo.User;
 import org.springframework.stereotype.Controller;
