+++ src/main/java/io/codeka/gaia/modules/controller/ModuleRestController.java
@@ -1,7 +1,7 @@
-package io.codeka.gaia.controller;
+package io.codeka.gaia.modules.controller;
 
-import io.codeka.gaia.bo.TerraformModule;
-import io.codeka.gaia.repository.TerraformModuleRepository;
+import io.codeka.gaia.modules.bo.TerraformModule;
+import io.codeka.gaia.modules.repository.TerraformModuleRepository;
 import io.codeka.gaia.teams.bo.User;
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.security.access.annotation.Secured;
