+++ src/main/java/io/codeka/gaia/modules/repository/TerraformCLIRepository.java
@@ -1,4 +1,4 @@
-package io.codeka.gaia.repository;
+package io.codeka.gaia.modules.repository;
 
 import org.apache.commons.lang.StringUtils;
 import org.springframework.beans.factory.annotation.Autowired;
