+++ src/main/java/io/codeka/gaia/modules/repository/TerraformModuleGitRepository.java
@@ -1,6 +1,6 @@
-package io.codeka.gaia.repository;
+package io.codeka.gaia.modules.repository;
 
-import io.codeka.gaia.bo.TerraformModule;
+import io.codeka.gaia.modules.bo.TerraformModule;
 import io.codeka.gaia.repository.strategy.GitHubStrategy;
 import io.codeka.gaia.repository.strategy.GitLabStrategy;
 import io.codeka.gaia.repository.strategy.GitPlatformStrategy;
