+++ src/main/java/io/codeka/gaia/modules/repository/strategy/GitHubStrategy.java
