+++ src/main/java/io/codeka/gaia/service/StackCostCalculator.java
@@ -4,7 +4,7 @@ import io.codeka.gaia.bo.JobStatus;
 import io.codeka.gaia.bo.JobType;
 import io.codeka.gaia.bo.Stack;
 import io.codeka.gaia.repository.JobRepository;
-import io.codeka.gaia.repository.TerraformModuleRepository;
+import io.codeka.gaia.modules.repository.TerraformModuleRepository;
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.stereotype.Service;
