+++ src/test/java/io/codeka/gaia/config/ActuatorSecurityTest.java
@@ -1,17 +1,10 @@
 package io.codeka.gaia.config;
 
-import io.codeka.gaia.repository.StackRepository;
-import io.codeka.gaia.repository.TerraformModuleRepository;
 import io.codeka.gaia.test.MongoContainer;
-import org.junit.jupiter.api.Disabled;
 import org.junit.jupiter.api.Test;
-import org.mockito.Mock;
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.boot.test.autoconfigure.web.servlet.AutoConfigureMockMvc;
-import org.springframework.boot.test.autoconfigure.web.servlet.WebMvcTest;
 import org.springframework.boot.test.context.SpringBootTest;
-import org.springframework.boot.test.mock.mockito.MockBean;
-import org.springframework.data.mongodb.core.MongoTemplate;
 import org.springframework.test.annotation.DirtiesContext;
 import org.springframework.test.web.servlet.MockMvc;
 import org.testcontainers.junit.jupiter.Container;
