+++ src/test/java/io/codeka/gaia/modules/repository/strategy/GitHubStrategyTest.java
