+++ src/test/java/io/codeka/gaia/modules/repository/strategy/GitLabStrategyTest.java
