+++ src/test/java/io/codeka/gaia/service/StackCostCalculatorTest.java
@@ -3,9 +3,9 @@ package io.codeka.gaia.service;
 import io.codeka.gaia.bo.Job;
 import io.codeka.gaia.bo.JobType;
 import io.codeka.gaia.bo.Stack;
-import io.codeka.gaia.bo.TerraformModule;
+import io.codeka.gaia.modules.bo.TerraformModule;
 import io.codeka.gaia.repository.JobRepository;
-import io.codeka.gaia.repository.TerraformModuleRepository;
+import io.codeka.gaia.modules.repository.TerraformModuleRepository;
 import io.codeka.gaia.teams.bo.User;
 import org.junit.jupiter.api.Test;
 import org.junit.jupiter.api.extension.ExtendWith;
